django_rest_api/requirements.txt
amqp==5.1.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==22.1.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
djando==0.2
Django==4.0.3
djangorestframework==3.13.1
kombu==5.2.4
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.0
prompt-toolkit==3.0.28
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
vine==5.0.0
wcwidth==0.2.5
